@@ -8,27 +8,21 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.2 <4.0.0'
+  sdk: '>=3.4.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  intl: ^0.18.0
-  file_saver: ^0.2.8
+  intl: ^0.19.0
 
-  flutter_datetime_picker: ^1.5.1
-  logger: ^2.0.2
-  flutter_local_notifications: ^15.1.1
-  timezone: ^0.9.2
-  flutter_timezone: ^1.0.7
-  rxdart: ^0.27.7
   # native
   file_picker: ^5.5.0
   url_launcher: ^6.1.14
-  widget_to_marker: ^1.0.4
   http_parser: ^4.0.2
+  flutter_local_notifications: ^15.1.1
+  file_saver: ^0.2.8
 
   # location
   google_maps_flutter: ^2.5.3
@@ -60,8 +54,13 @@ dependencies:
   http: ^1.1.0 # todo: remove
 
   # integrations
+  rxdart: ^0.27.7
   flutter_svg: ^2.0.7
   cached_network_image: ^3.3.1
+  flutter_to_pdf: ^0.2.1
+  widget_to_marker: ^1.0.4
+  timezone: ^0.9.2
+  flutter_timezone: ^1.0.7
 
   # icons
   fluentui_system_icons: ^1.1.234
@@ -74,8 +73,9 @@ dependencies:
   dropdown_button2: ^2.3.9
   dotted_border: ^2.1.0
   extended_text: ^13.0.0
-
+  
   # cicd
+  logger: ^2.3.0
   sentry_flutter: ^7.14.0
   device_info_plus: ^9.1.2
   flutter_background_service: ^5.0.5
@@ -101,12 +101,20 @@ dev_dependencies:
   flutter_gen_runner: ^5.3.1
   json_serializable: ^6.7.1
   go_router_builder: ^2.3.2
+
+  #* todo: oz remove this
+  flutter_datetime_picker: ^1.5.1
+  
 dependency_overrides:
   flutter_gen_core:
     git:
       url: https://github.com/FlutterGen/flutter_gen.git
       ref: 8c57b5aa73f7793d614f064ec8e00434cc4cf01a
       path: packages/core
+  kosher_dart:
+    git:
+      url: https://github.com/noga-dev/kosher_dart.git
+      ref: master
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
